Add files via upload
It now reads and writes events to and from firestore and allows for deletion of events.  Hoping to get editing working today as well.  From then on it will be functional and it will be more or less just quality of life changes that need to be changed (updates to ui etc).